refactor: move cytoscape stylesheet to topologyStyle.ts
